Remove React component from `ts` import to fix tests
